requirements.txt
pip>20.0.0
black==20.8b1

requests==2.24.0
python-dotenv==0.14
SQLAlchemy==1.3.19
PyMySQL==0.10.0

passlib==1.7.2
fastapi==0.61.1
uvicorn==0.11.8
gunicorn==20.0.4
pytz==2020.1
